WittVectors.jl: An implementation of big Witt vectors supporting truncation over arbitrary divisor-stable sets

WittVectors.jl: An implementation of big Witt vectors supporting truncation over arbitrary divisor-stable sets

Parent and child objects and constructors

Data type and parent object methods

Basic Manipulation

Binary Operations

Exact Division

Comparison

Truncation Helpers

Parent and child objects and constructors

WittVectors.BigWittRing — Type

mutable struct BigWittRing{T <: RingElement} <: Ring

Parent object type for Big Witt Rings (i.e. truncated only by a maximum precision rather than a more general divisor-stable set). Should be constructed using the exported constructors, although I think it's basically safe to call WittVectors.BigWittRing directly.

(W::BigWittRing{T})() where T <: RingElement

The basic constructor given a Witt Ring W. Identical to zero(W) (in fact the latter simply calls W())

Example:

julia> using AbstractAlgebra;

julia> using WittVectors;

julia> W=BigWittVectorRing(ZZ,10)
Big Witt vector ring represented up to degree 10 over Integers

julia> w=W()
BigInt[0, 0, 0, 0, 0, 0, 0, 0, 0, 0]

julia> iszero(w)
true

Constructors

(W::BigWittRing{T})(c::Integer) where T <: RingElement
(W::BigWittRing{T})(c::T) where T <: RingElement

Since there is no additive ring homomorphism R→W(R), calling W(c) for c an element of R returns the constant lift of c. This is more-or-less the only candidate for that functionality, but for that reason W(c) for c an integer does not behave the way it does for the other functorial constructions out of the category of Rings of AbstractAlgebra.jl. All of the pre-existing such $F$ (to my knowledge) admit a natural transformation $id ⟹ F$, so in those cases calling $FR(c)$ where c may be interpreted as either an element of R or as a Julia integral type is unambiguous and returns the image of c in the composition of structure maps $Z→R→FR$. Since we do not have such a natural transformation at our disposal, to avoid ambiguity we have defined W(c) to be the constant lift of the image of c in $ZZ→R$.

Example:

julia> using WittVectors;

julia> using AbstractAlgebra;

julia> W1=BigWittVectorRing(ZZ,10)
Big Witt vector ring represented up to degree 10 over Integers

julia> W1(256)
BigInt[256, 0, 0, 0, 0, 0, 0, 0, 0, 0]

julia> W2=BigWittVectorRing(GF(7),10)
Big Witt vector ring represented up to degree 10 over Finite field F_7

julia> W2(256)
AbstractAlgebra.GFElem{Int64}[4, 0, 0, 0, 0, 0, 0, 0, 0, 0]

julia> W3=BigWittVectorRing(AbsSeriesRing(GF(23),10),10)
Big Witt vector ring represented up to degree 10 over Univariate power series ring in x over Finite field F_23

julia> W3(256)
AbstractAlgebra.Generic.AbsSeries{AbstractAlgebra.GFElem{Int64}}[3 + O(x^10), O(x^10), O(x^10), O(x^10), O(x^10), O(x^10), O(x^10), O(x^10), O(x^10), O(x^10)]

The second one is possibly needed to prevent ambiguity, according to the ring interface documentation for AbstractAlgebra.jl.

(W::BigWittRing{T})(c::T) where T <: RingElement

This is the constant lift $R→W(R)$, which is multiplicative but not additive.

DocTestSetup = quote
	using AbstractAlgebra, WittVectors, Hecke
end

julia> using AbstractAlgebra, Hecke, WittVectors, Nemo;

julia> Qx,x = PolynomialRing(Hecke.QQ, "x")
(Univariate Polynomial Ring in x over Rational Field, x)

julia> f=x^3+x+1
x^3 + x + 1

julia> R, a=Hecke.NumberField(f, 'a', cached=true, check=true)
(Number field over Rational Field with defining polynomial x^3 + x + 1, a)

julia> W=BigWittVectorRing(R,8)
Big Witt vector ring represented up to degree 8 over Number field over Rational Field with defining polynomial x^3 + x + 1

julia> W(a+2)
Nemo.nf_elem[a + 2, 0, 0, 0, 0, 0, 0, 0]

DocTestSetup = nothing

(W::BigWittRing{T})(A::Vector{T}) where T <: RingElem

Additional constructor to create a Witt vector by supplying its coordinates. Will return an error if supplied more coordinates than W.prec

WittVectors.WittVector — Type

mutable struct WittVector{T <: RingElement} <: RingElem

Object type for Big Witt Vectors, i.e. child objcets of BigWittRing{T}s. DO NOT CONSTRUCT USING WittVectors.WittVector{T}(coords, prec), this will produce an orphan Witt Vector which you would then have to assign a parent to directly. Parent-child compatibility is geneerally unchecked, so no lifeguard on duty. Rather, construct with constructors below.

WittVectors.BigWittVectorRing — Function

BigWittVectorRing(R::Ring, n::Integer, cached::Bool=true)

Main constructor for creating rings of Witt vectors. Returns parent object for the ring of Witt vectors over the base ring R up to degree n. The optional cached argument determines whether to cache the returned object, so that calling the same constructor with the same arguments R,n twice returns the same object and hence child objects will be compatible–no life guard on duty if you call BigWittVectorRing with cached==false.

WittVectors.TruncatedBigWittRing — Type

mutable struct TruncatedBigWittRing{T <: RingElement} <: Ring

Parent object for truncated rings of Witt vectors. Contains two fields, untruncated::BigWittRing{T} which is the underlying ring (prior to quotient) and truncationset::Vector{Bool} which determines which indices are kept in the truncation. truncationset should be the same length as untruncated.prec.

Constructors

function (W::TruncatedBigWittRing{T})() where T <: RingElement
function (W::TruncatedBigWittRing{T})(c::Integer) where T <: RingElement
function (W::TruncatedBigWittRing{T})(c::T) where T <: Integer
function (W::TruncatedBigWittRing{T})(c::T) where T <: RingElement

Each of these mirror the functionality in the untruncated case.

function (W::TruncatedBigWittRing{T})(A::Vector{T}) where T <: RingElem

Return a Witt vector with coordinates specified by A. A can be the length of W.truncationset (i.e. W.untruncated.prec in which case it will ignore the entries of A ignored by W or it can be length equal to the number of true entries in W.truncationset, in which case it will fill in the missing (irrelevant) entries with zero.

WittVectors.TruncatedWittVector — Type

mutable struct TruncatedWittVector{T <: RingElement} <: RingElem

Child type for truncated Witt vectors. Contains three fields, xcoords::Vector{T} which are its coordinates (INCLUDING indices not in the truncation set), truncationset::Vector{Bool} and parent::TruncatedBigWittRing{T}. truncationset and parent.truncationset should match.

WittVectors.TruncatedBigWittVectorRing — Function

function TruncatedBigWittVectorRing(R::Ring, S::Vector{Bool}, cached::Bool=true, stabilize::Bool=true)

Return a ring of Witt vectors over the base ring R truncated by the truncation set S.

Similarly to the corresponding untruncated constructor, the cached argument determines whether to add the created ring to the parent types dictionary. Recommended this is always true unless you really know what you are doing. The stabilize option determines whether to replace the truncation set S with divisor_stabilize(S). One can expect a slight speedup if called with stabilize=false the first time the ring is created, but this will induce unexpected behavior (namely it will in all likelihood break the associativity of the + and * operators) if S is not divisor-stable to begin with. Recommended you call with stabilize=true (the default option if no argument specified) unless you are creating a great number of Witt rings at once with truncation sets you know to be divisor-stable and need the speedup. The resulting object W will have W.untruncated.prec equal to the length of the supplied truncation set. Recommended the last entry of S is true in order to optimize performance and decrease the chances of creating functionally identical parent objects which are unequal.

function TruncatedBigWittVectorRing(R::Ring, Ns::Vector{Int}, cached::Bool=true, stabilize::Bool=true)

Same as above except now taking an integer-vector argument.

function TruncatedBigWittVectorRing(W::BigWittRing, S::Vector{Bool}, cached::Bool=true, stabilize::Bool=true)

Return TruncatedBigWittVectorRing(W.base_ring, S, cached, stabilize).

WittVectors.pTypicalWittVectorRing — Function

function pTypicalWittVectorRing(R::Ring, p::Integer, m::Integer,cached::Bool=true)

Specialized version of the TruncatedBigWittVectorRing to create $W_{p^m}(R)$, the ring of p-typical Witt vectors represented up to entry p^m. Resulting object W will have W.untruncated.prec equal to p^m.

pTypicalWittVectorRing(W::BigWittRing, p::Integer, m::Integer, cached::Bool=true)

Return pTypicalWittVectorRing(base_ring(W), p, m, cached)

Data type and parent object methods

AbstractAlgebra.divexact — Method

divexact(w::WittVector{T}, v::WittVector{T}; check::Bool=true) where T <: RingElement
inv(w::WittVector)

These are effectively not implemented–it has exactly what it needs to in order not to fail the conformance tests, which is to say that if v=W(1), divexact(w,v) returns w. Otherwise, it returns an error. Like everything else having to do with units in the ring of Big Witt Vectors, this may be expanded eventually, but no promises.

Basic Manipulation

Base.zero — Function

zero(W::BigWittRing)
one(W::BigWittRing)
isone(w::WittVector)
iszero(w::WittVector)

More standard ring interface functions. Check the documentation for AbstractAlgebra.jl's generic ring interface

AbstractAlgebra.isconstant — Function

function isconstant(w::WittVector)

Returns true if w is in the image of the constant lift set map R→W(R) (better known by another name, but we are avoiding the Nazi associations (somewhat unavoidable in the case of Ernst Witt)).

AbstractAlgebra.is_unit — Method

is_unit(w::WittVector)

Returns true if w is known to be a unit. Known being the operative word here–outside of special cases like W_p(Z/p), I do not really know a good way to know whether a given Witt vector is a unit. This functionality may be expanded in the future (but perhaps do not hold your breath).

Binary Operations

Base.:* — Function

Z-algebra Structure

*(w::WittVector{T}, c::Integer) where T <: RingElement

Since there is no additive ring homomorphism R→W(R), calling W(c) for c an element of R returns the constant lift of c. This is more-or-less the only candidate for that functionality, but for that reason W(c) for c an integer does not behave the way it does for the other functorial constructions out of the category of Rings of AbstractAlgebra.jl. All of the pre-existing such F (to my knowledge) admit a natural transformation id ⟹F, so in those cases calling FR(c) where c may be interpreted as either an element of R or as a Julia integral type is unambiguous and returns the image of c in the composition of structure maps Z→R→FR. Since we do not have such a natural transformation at our disposal, to avoid ambiguity we have defined W(c) to be the constant lift of the image of c in Z→R. Thus, in particular, for a Witt Vector w∈W, W(2)*w ≂̸ w+w (again, unlike other functorial ring constructions in Julia). To compensate, an additional method has been added to the multiplication function * to accept mixed arguments.

Example:

julia> using AbstractAlgebra;

julia> using WittVectors;

julia> W=BigWittVectorRing(ZZ,10)
Big Witt vector ring represented up to degree 10 over Integers

julia> w=one(W)
BigInt[1, 0, 0, 0, 0, 0, 0, 0, 0, 0]

julia> x=W(2)*w
BigInt[2, 0, 0, 0, 0, 0, 0, 0, 0, 0]

julia> y=2*w
BigInt[2, -1, -2, -4, -6, -12, -18, -40, -54, -120]

julia> x==w+w
false

julia> y==w+w
true

julia> z=w*2
BigInt[2, -1, -2, -4, -6, -12, -18, -40, -54, -120]

julia> z==w+w
true

julia> t=(-2)*w
BigInt[-2, -3, 2, -9, 6, -4, 18, -93, 54, -72]

julia> t + w + w == 0
true

Exact Division

AbstractAlgebra.parent_type — Method

Data type and parent object methods

parent_type(::Type{WittVector{T}}) where T <: RingElement
elem_type (::Type{WittVector{T}}) where T <: RingElement
base_ring(R::BigWittRing)
parent(R::WittVector)
is_domain_type(::Type{WittVector{T}}) where T <: RingElement
is_exact_type(::Type{WittVector{T}}) where T <: RingElement
hash(w::WittVector, h::UInt)
deepcopy_internal(w::WittVector{T}, d::IdDict) where T <: RingElement

Standard components of any conformant ring implementation in AbstractAlgebra.jl. See the documentation there

Comparison

Base.:== — Function

function ==(w::TruncatedWittVector{T}, v::TruncatedWittVector{T}) where T<: RingElement

Return true if and only if w and v have the same parent and w[i]==v[i] for i ∈ w.truncationset (equivelently i ∈ v.truncationset). In general any TruncatedWittVector constructed normally will have zeros for those entries anyway, but in the event this property is somehow broken, that will be invisible to ==.

Truncation Helpers

WittVectors.truncationbools — Function

function truncationbools(X::Vector{Int},n::Int)

Return vector of booleans corresponding to a truncationset given as a Vector, Array or Matrix of integers. Unexpected results may occur if supplied a matrix with both dimensions exceeding 1. The second argument n specifies the length of the returned vector. n should exceed the largest member of X, will throw an index error otherwise.

function truncationbools(X)

Alias for trunationbools(X,n) where n is the maximal member of X.

WittVectors.truncationlist — Function

function truncationlist(S::Vector{Bool})

Return a list of integers representing the truncationset given its representation as a boolean vector.

WittVectors.divisor_stabilize — Function

function divisor_stabilize(S::Vector{Bool})

Return a divisor-stable truncationset (as a Vector{Bool}) generated by its argument. For now a very slow and naive implementation, but hardly performance-critical.

Example

julia> using WittVectors;

julia> using AbstractAlgebra;

julia> S=zeros(Bool, 36);

julia> S[36]=true
true

julia> S
36-element Vector{Bool}:
 0
 0
 0
 0
 0
 0
 0
 0
 0
 0
 ⋮
 0
 0
 0
 0
 0
 0
 0
 0
 1

julia> divisor_stabilize(S)
36-element Vector{Bool}:
 1
 1
 1
 1
 0
 1
 0
 0
 1
 0
 ⋮
 0
 0
 0
 0
 0
 0
 0
 0
 1

Base.truncate — Function

truncate(W::BigWittRing, S::Vector{Bool})

Return a TruncatedBigWittVectorRing W_new with W_new.untruncated being W, truncated by S.

function truncate(w::WittVector, S::Vector{Bool})

Return the image of w in the quotient map $W(R)→W_S(R)$. Parent object will be truncate(w.parent, S)

WittVectors.BigWittRing

WittVectors.TruncatedBigWittRing

WittVectors.TruncatedWittVector

WittVectors.WittVector

AbstractAlgebra.divexact

AbstractAlgebra.is_unit

AbstractAlgebra.isconstant

AbstractAlgebra.parent_type

Base.:*

Base.:==

Base.truncate

Base.zero

WittVectors.BigWittVectorRing

WittVectors.TruncatedBigWittVectorRing

WittVectors.divisor_stabilize

WittVectors.pTypicalWittVectorRing

WittVectors.truncationbools

WittVectors.truncationlist